test: @DataRedisTest를 이용하여 redis repository test code 작성
